Mobile improvements
This is following the commit last night where I fixed the headers on mobile. This code is no longer "dry" but at this point I think it is fine because if I need to change the nav or the footer I know exactly where it is on the html so I can just copy and paste the new version in.