It happened...
I have decided after trying to fix this bug to just throw in the towel. I don't like doing it but I have started to realise that it is more complicated to maintain with this background on the index page and there are more bugs that I can't seem to fix fast enough.